requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19

# Document processing
PyMuPDF==1.25.1
pdfplumber==0.11.4
python-docx==1.1.2
pytesseract==0.3.13
Pillow==11.0.0
openpyxl==3.1.5

# Embeddings and vector store
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
chromadb==0.5.23

# Gemini API
google-generativeai==0.8.3

# Utilities
python-dotenv==1.0.1
tiktoken==0.8.0
aiofiles==24.1.0
pydantic==2.10.3
pydantic-settings==2.7.0
